+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPR03Y7HSA6SMF36TQJTV1KQ"
+        "01KPR1AQT8EWEWYK0GNXDCKE4V"
     )
     r"""The unique identifier of the data source"""