+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPR03Y7GDRVF05JMBA18X234"
+        "01KPR1AQT7J7GMH9FXZ7YCHSVA"
     )
     r"""The unique identifier of the data source"""