+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPR03Y7KMWYJNPKNM5NJF554"
+        "01KPR1AQTAWC0X7W14Y257ZK8C"
     )
     r"""The unique identifier of the data source"""